rename method: givenUserWithExistingName_whenSaveUser_thenGiveUsernameAlreadyExistsError
rename this class according to the standard -- UserServiceUnitTest.java
